requirements.txt
beautifulsoup4
cloudscraper
lxml
Pillow
pypdf
requests
pywidevine
cryptography
